Make sure job is fully saved before callback is called
Previously the logic called the callback before all state related to the job was saved (the rest was saved in the background).  This works in many cases, but when run from a command line process which creates a job and then quits, jobs were not getting created.

This change makes sure that the job is fully saved before the callback is called.