Be more tolerant of bad job data
This fix lets the UI be more tolerant of null fields in jobs.
While these should not typically occur, I have seen cases where a job ends up with null fields due to redis bug/corruption issues.
Once the job is corrupted, it won't show up in the UI (so you can't delete it, fix it, etc. without doing so manually).
This change lets the UI be a bit more lenient so corrupt jobs can still be seen and deleted.