Handle errors emitted by redis by listening on error event to avoid crashes
http://nodejs.org/api/events.html#events_class_events_eventemitter
When an EventEmitter instance experiences an error, the typical action is to emit an 'error' event. Error events are treated as a special case in node. If there is no listener for it, then the default action is to print a stack trace and exit the program.
Redis emits errors in several places.