Fixing FIFO problem
Redis sorts zsets lexographically for members with the same priority.
This causes jobs to run out of order. If you submit 11 jobs at the same
priority level, they will run as 1,10,11,2,3,4,5,6,7,8,9.

This adds a zid to the job that prepends the number of digits in the job
id (001)1. This new zset value will be sorted correctly by redis. This
is not desirable as a general purpose id, so it is hidden and the
standard jobid can be used for all external purposes.

The prepended value is stripped off when doing zpop.