Added main.css to the repo (security)
As described in #713 - https://github.com/Automattic/kue/issues/713

node generates main.css out of the styles files on demand. This might be convenient for many scenarios, but this is a big security issue because node must have write access to its own code directory -- thus some bug might allow attacker to overwrite ones own code. IMO, Kue should include main.css in the repository, so that a locked-down server without write access wouldn't break it.

To regenerate main.css, simply delete it locally and run the code that shows kue site. Once you navigate to it in a browser, the main.css will be regenerated, and can be updated in git.