new method #unique() prevent dup jobs
The method can be used with (unique on specific field) or without
(unique on all data) arguments.

When used, Kue will not create new jobs with the same unique field.

We keep track of the unique fields with an HMAP,
mapping a uniqueKey => jobId. We optimistically set the mapping
(producing churn in the incrementing jobId value), but do **not** create
a new job if we detect the mapping already existed.

Note that the use of #removeOnComplete will allow new jobs with the same
unique field(s) to be created. This is often desireable.